Update to version 3.2.10
Change the word line to the builtin REPLY variable of the read builtin.
Replace echo with printf when printing Missing message inside the loop.
Remove the negation in checking the value of ${#Missing[@]}.
  * Change the || into && before the die function.
  * Add a 127 exit status. (the exit status of cnf).
Add a variable that holds some fancy format when printing changes.
Use direct command susbstituion when printing number of lines. (no more variables)
Update the readme file
  * Clarify if and when the crontab entry is going to be created.
  * Add instruction how to download and install and use HistLog. (system wide)